requirements.txt
azure-search-documents
azure-core
openai
python-dotenv